requirements.txt
python-kucoin>=2.1.3
kucoin-futures-python>=1.0.6
selenium>=4.4.3
python-telegram-bot>=13.14
pyyaml